Function SendMail Finished

+++ src/main/java/mg/meddoc/utils/Util.java
@@ -7,7 +7,7 @@ public class Util {
 	public static String generateCode() {
 		String code = "";
 		for(int i = 0; i < 6; i++) {
-			code+=(Math.random()*10);
+			code+=(int)(Math.random()*10);
 		}
 		return code;
 	}
